0.1.1
fatfs 32bit->64bit

+++ FATFS/src/integer.h
@@ -27,8 +27,8 @@ typedef unsigned short	WORD;
 typedef unsigned short	WCHAR;
 
 /* These types MUST be 32-bit */
-typedef long			LONG;
-typedef unsigned long	DWORD;
+typedef long long			LONG;
+typedef unsigned long long	DWORD;
 
 /* This type MUST be 64-bit (Remove this for ANSI C (C89) compatibility) */
 typedef unsigned long long QWORD;
